fixed overflow problems

+++ atcoder/abc161/d.cpp
@@ -63,11 +63,11 @@ ll modpow(ll a, ll n) {
 int main() {
     cin.tie(nullptr);
     ios::sync_with_stdio(false);
-    int k; cin >> k;
-    queue<int> q;
+    ll k; cin >> k;
+    queue<ll> q;
     rep1(i, 9) q.push(i);
     rep(i, k-1) {
-        int x = q.front(); q.pop();
+        ll x = q.front(); q.pop();
         if (x % 10 != 0) q.push(10 * x + (x % 10) - 1);
         q.push(10 * x + (x % 10));
         if (x % 10 != 9) q.push(10 * x + (x % 10) + 1);
